+++ abstractions/go/serialization/parsable.go
@@ -1,0 +1,6 @@
+package serialization
+
+type Parsable interface {
+	Serialize(writer SerializationWriter) error
+	GetFieldDeserializers() map[string]func(interface{}, ParseNode) error
+}