+++ src/Kiota.Builder/GenerationConfiguration.cs
@@ -13,5 +13,18 @@ namespace Kiota.Builder {
         public bool UsesBackingStore { get; set; }
         public List<string> Serializers { get; set; } = new();
         public List<string> Deserializers { get; set; } = new();
+        public bool ShouldWriteNamespaceIndices { get { return BarreledLanguages.Contains(Language); } }
+        public bool ShouldWriteBarrelsIfClassExists { get { return BarreledLanguagesWithConstantFileName.Contains(Language); } }
+        public bool ShouldRenderMethodsOutsideOfClasses { get { return MethodOutsideOfClassesLanguages.Contains(Language); } }
+        private static HashSet<GenerationLanguage> MethodOutsideOfClassesLanguages = new () {
+            GenerationLanguage.Go,
+        };
+        private static HashSet<GenerationLanguage> BarreledLanguages = new () {
+            GenerationLanguage.Ruby,
+            // TODO: add typescript once we have a barrel writer for it
+        };
+        private static HashSet<GenerationLanguage> BarreledLanguagesWithConstantFileName = new () {
+            //TODO: add typescript once we have a barrel writer for it
+        };
     }
 }