+++ src/Kiota.Builder/PathSegmenters/CSharpPathSegmenter.cs
@@ -7,6 +7,6 @@ namespace Kiota.Builder {
         public CSharpPathSegmenter(string rootPath, string clientNamespaceName): base(rootPath, clientNamespaceName) { }
         public override string FileSuffix => ".cs";
         public override string NormalizeNamespaceSegment(string segmentName) => segmentName.ToFirstCharacterUpperCase();
-        public override string NormalizeFileName(string elementName) => elementName.ToFirstCharacterUpperCase();
+        public override string NormalizeFileName(CodeElement currentElement) => GetLastFileNameSegment(currentElement).ToFirstCharacterUpperCase();
     }
 }