+++ src/Kiota.Builder/Refiners/JavaRefiner.cs
@@ -9,16 +9,16 @@ namespace Kiota.Builder.Refiners {
         public JavaRefiner(GenerationConfiguration configuration) : base(configuration) {}
         public override void Refine(CodeNamespace generatedCode)
         {
-            AddInnerClasses(generatedCode);
+            AddInnerClasses(generatedCode, false);
             AndInsertOverrideMethodForRequestExecutorsAndBuilders(generatedCode);
             ReplaceIndexersByMethodsWithParameter(generatedCode, generatedCode);
             ConvertUnionTypesToWrapper(generatedCode);
             AddRequireNonNullImports(generatedCode);
             FixReferencesToEntityType(generatedCode);
             AddPropertiesAndMethodTypesImports(generatedCode, true, false, true);
-            AddDefaultImports(generatedCode, defaultNamespaces, defaultNamespacesForModels, defaultNamespacesForRequestBuilders, defaultSymbolsForApiClient);
+            AddDefaultImports(generatedCode, defaultNamespaces, defaultNamespacesForModels, defaultNamespacesForRequestBuilders);
             CorrectCoreType(generatedCode, CorrectMethodType, CorrectPropertyType);
-            PatchHeaderParametersType(generatedCode);
+            PatchHeaderParametersType(generatedCode, "Map<String, String>");
             AddListImport(generatedCode);
             AddParsableInheritanceForModelClasses(generatedCode);
             ReplaceBinaryByNativeType(generatedCode, "InputStream", "java.io", true);
@@ -33,7 +33,10 @@ namespace Kiota.Builder.Refiners {
             CorrectCoreTypesForBackingStore(generatedCode, "com.microsoft.kiota.store", "BackingStoreFactorySingleton.instance.createBackingStore()");
             ReplaceDefaultSerializationModules(generatedCode, "com.microsoft.kiota.serialization.JsonSerializationWriterFactory");
             ReplaceDefaultDeserializationModules(generatedCode, "com.microsoft.kiota.serialization.JsonParseNodeFactory");
-            AddSerializationModulesImport(generatedCode);
+            AddSerializationModulesImport(generatedCode,
+                                        new [] { "com.microsoft.kiota.ApiClientBuilder",
+                                                "com.microsoft.kiota.serialization.SerializationWriterFactoryRegistry" },
+                                        new [] { "com.microsoft.kiota.serialization.ParseNodeFactoryRegistry" });
         }
         private static void AddEnumSetImport(CodeElement currentElement) {
             if(currentElement is CodeClass currentClass && currentClass.IsOfKind(CodeClassKind.Model) &&
@@ -95,11 +98,6 @@ namespace Kiota.Builder.Refiners {
             new ("BiConsumer", "java.util.function"),
             new ("Map", "java.util"),
             new ("HashMap", "java.util"),
-        };
-        private static readonly Tuple<string, string>[] defaultSymbolsForApiClient = new Tuple<string, string>[] { 
-            new ("ApiClientBuilder", "com.microsoft.kiota"),
-            new ("SerializationWriterFactoryRegistry", "com.microsoft.kiota.serialization"),
-            new ("ParseNodeFactoryRegistry", "com.microsoft.kiota.serialization"),
         };
         private static void CorrectPropertyType(CodeProperty currentProperty) {
             if(currentProperty.IsOfKind(CodePropertyKind.HttpCore))
@@ -181,13 +179,6 @@ namespace Kiota.Builder.Refiners {
             }
             
             CrawlTree(currentElement, AndInsertOverrideMethodForRequestExecutorsAndBuilders);
-        }
-        private static void PatchHeaderParametersType(CodeElement currentElement) {
-            if(currentElement is CodeMethod currentMethod && currentMethod.Parameters.Any(x => x.IsOfKind(CodeParameterKind.Headers)))
-                currentMethod.Parameters.Where(x => x.IsOfKind(CodeParameterKind.Headers))
-                                        .ToList()
-                                        .ForEach(x => x.Type.Name = "Map<String, String>");
-            CrawlTree(currentElement, PatchHeaderParametersType);
         }
         private static CodeMethod GetMethodClone(CodeMethod currentMethod, params CodeParameterKind[] parameterTypesToExclude) {
             if(currentMethod.Parameters.Any(x => parameterTypesToExclude.Contains(x.ParameterKind))) {