+++ src/Kiota.Builder/Writers/TypeScript/CodeMethodWriter.cs
@@ -63,6 +63,8 @@ namespace Kiota.Builder.Writers.TypeScript {
                 case CodeMethodKind.Constructor:
                     WriteConstructorBody(parentClass, codeElement, writer, inherits);
                     break;
+                case CodeMethodKind.RequestBuilderBackwardCompatibility:
+                    throw new InvalidOperationException("RequestBuilderBackwardCompatibility is not supported as the request builders are implemented by properties.");
                 default:
                     WriteDefaultMethodBody(codeElement, writer);
                     break;