+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ruby JSON serialization #429
 - Ruby HTTP service #472
+- Go generation support & abstractions #413
 
 ### Changed